Move 'FeedViewAdapter' to separeate file
